requirements.txt
django==4.2.1
psycopg2==2.9.9
pillow==10.4.0
djangorestframework==3.15.2
